fix log quality record

+++ scripts/img2video/main.py
@@ -128,6 +128,9 @@ class Image2VideoWebsocket(tornado.websocket.WebSocketHandler):
 
     def on_close(self):
         self.writer.close()
+        if not os.path.exists(self.video_tmp_path):
+            print("no video file generated")
+            return
         shutil.move(self.video_tmp_path, self.video_path)
         with open(self.video_path + '.json', 'wb') as f:
             f.write(
@@ -135,7 +138,7 @@ class Image2VideoWebsocket(tornado.websocket.WebSocketHandler):
                     "udid": self.udid,
                     "name": self.name
                 }).encode('utf-8'))
-        print("websocket closed, video genreated", self.video_path)
+        print("websocket closed, video generated", self.video_path)
 
 
 class Image2VideoHandler(CorsMixin, tornado.web.RequestHandler):
